requirements.txt
Flask
requests
mysql-connector-python
APScheduler
